AutoSprayonator Fix
