+++ gradle.properties
@@ -4,5 +4,5 @@ baseGroup=com.jelly.farmhelperv2
 mcVersion=1.8.9
 modid=farmhelperv2
 modName=FarmHelper
-version=2.2.7
+version=2.2.9
 shouldRelease=true
