+++ src/main/java/com/jelly/farmhelperv2/feature/FeatureManager.java
@@ -36,7 +36,6 @@ public class FeatureManager {
             PerformanceMode.getInstance(),
             PestsDestroyer.getInstance(),
             AutoSprayonator.getInstance(),
-            NewAutoSprayonator.getInstance(),
             PetSwapper.getInstance(),
             PlotCleaningHelper.getInstance(),
             ProfitCalculator.getInstance(),
