+++ src/main/java/com/jelly/farmhelperv2/feature/impl/AutoSprayonator.java
@@ -1,18 +1,12 @@
 package com.jelly.farmhelperv2.feature.impl;
 
-import cc.polyfrost.oneconfig.utils.Multithreading;
 import com.jelly.farmhelperv2.config.FarmHelperConfig;
-import com.jelly.farmhelperv2.config.FarmHelperConfig.SPRAYONATOR_ITEM;
-import com.jelly.farmhelperv2.event.DrawScreenAfterEvent;
 import com.jelly.farmhelperv2.feature.FeatureManager;
 import com.jelly.farmhelperv2.feature.IFeature;
 import com.jelly.farmhelperv2.handler.GameStateHandler;
 import com.jelly.farmhelperv2.handler.MacroHandler;
 import com.jelly.farmhelperv2.handler.RotationHandler;
-import com.jelly.farmhelperv2.util.InventoryUtils;
-import com.jelly.farmhelperv2.util.KeyBindUtils;
-import com.jelly.farmhelperv2.util.LogUtils;
-import com.jelly.farmhelperv2.util.PlotUtils;
+import com.jelly.farmhelperv2.util.*;
 import com.jelly.farmhelperv2.util.helper.Clock;
 import com.jelly.farmhelperv2.util.helper.Rotation;
 import com.jelly.farmhelperv2.util.helper.RotationConfiguration;
@@ -20,58 +14,52 @@ import lombok.Getter;
 import lombok.Setter;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.inventory.GuiChest;
-import net.minecraft.client.gui.inventory.GuiInventory;
 import net.minecraft.inventory.ContainerChest;
 import net.minecraft.inventory.Slot;
+import net.minecraft.item.ItemStack;
 import net.minecraft.util.StringUtils;
 import net.minecraftforge.client.event.ClientChatReceivedEvent;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.fml.common.gameevent.TickEvent;
 
-import java.util.HashMap;
-import java.util.List;
-import java.util.concurrent.TimeUnit;
+import java.util.Arrays;
+import java.util.Objects;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class AutoSprayonator implements IFeature {
-    @Getter
-    private static final AutoSprayonator instance = new AutoSprayonator();
+    private static AutoSprayonator instance = null;
 
-    @Getter
-    private final HashMap<Integer, PlotData> sprayonatorPlotStates = new HashMap<>();
+    public static AutoSprayonator getInstance() {
+        if (instance == null) {
+            instance = new AutoSprayonator();
+        }
+        return instance;
+    }
+
     private final Minecraft mc = Minecraft.getMinecraft();
-    private final String skymartItemName = "Sprayonator";
-    @Getter
-    private final Clock sprayonatorDelay = new Clock();
-    private boolean running;
-    @Getter
-    private SPRAYONATOR_ITEM sprayItem = SPRAYONATOR_ITEM.values()[FarmHelperConfig.sprayonatorType];
-    @Getter
-    private AUTO_SPRAYONATOR_STATE sprayState = AUTO_SPRAYONATOR_STATE.NONE;
-    @Getter
-    private CHECK_PLOT_STATE checkPlotState = CHECK_PLOT_STATE.NONE;
-    @Getter
-    private SKYMART_PURCHASE_STATE skymartPurchaseState = SKYMART_PURCHASE_STATE.NONE;
-    @Getter
-    private BAZAAR_PURCHASE_STATE bazaarPurchaseState = BAZAAR_PURCHASE_STATE.NONE;
-    @Getter
-    private CURRENT_GUI_STATE currentGuiState = CURRENT_GUI_STATE.NONE;
-    private final Pattern sprayTimerPattern = Pattern.compile("([a-zA-Z\\s]*)\\s?(?:(\\d+)m)?\\s+(\\d+)s");
-    private boolean hasCopper = true;
-    private String bazaarItemName;
-
-    private float[] prevRotation = new float[]{0f,0f};
-    private int rotationState = 0;
+    private final String SPRAYONATOR = "Sprayonator";
+    private boolean enabled = false;
+    private MainState mainState = MainState.NONE;
+    private CheckPlotState checkState = CheckPlotState.STARTING;
+    private BuyState buyState = BuyState.STARTING;
+    private SprayState sprayState = SprayState.STARTING;
+    private final String[] SPRAYONATOR_ITEM = {"Compost", "Honey Jar", "Dung", "Plant Matter", "Tasty Cheese"};
+    private Plot[] plots = new Plot[25];
+    private final Clock timer = new Clock();
+    private final Pattern sprayonatorTimePattern = Pattern.compile("(((\\d+)m\\s)?(\\d+)s)");
+    private int counter = 0; // I literally forgot the proper word kill me
+    private boolean isRotating = false;
+    private boolean changedMaterial = false;
 
     @Override
     public String getName() {
-        return "Auto Sprayonator";
+        return "AutoSprayonator";
     }
 
     @Override
     public boolean isRunning() {
-        return running;
+        return this.enabled;
     }
 
     @Override
@@ -81,53 +69,50 @@ public class AutoSprayonator implements IFeature {
 
     @Override
     public boolean shouldStartAtMacroStart() {
-        return true;
+        return false;
     }
 
     @Override
     public void start() {
-        if (running) return;
-        MacroHandler.getInstance().pauseMacro();
-        running = true;
+        if (this.enabled) return;
 
-        sprayItem = SPRAYONATOR_ITEM.values()[FarmHelperConfig.sprayonatorType];
-        if (sprayonatorPlotStates.isEmpty()) {
-            sprayState = AUTO_SPRAYONATOR_STATE.CHECK_PLOTS;
-        } else {
-            sprayState = AUTO_SPRAYONATOR_STATE.WAITING_FOR_PLOT;
+        if (!InventoryUtils.hasItemInInventory(SPRAYONATOR)) {
+            LogUtils.sendError("Cannot Find Sprayonator in Inventory. Disabling AutoSprayonator.");
+            FarmHelperConfig.autoSprayonatorEnable = false;
+            return;
         }
-        sprayonatorDelay.schedule(100);
+
+        this.enabled = true;
+        this.mainState = MainState.STARTING;
+        MacroHandler.getInstance().pauseMacro();
+        log("Activating");
     }
 
     @Override
     public void stop() {
-        if (!running) return;
-        LogUtils.sendWarning("[Auto Sprayonator] Stopping");
-        resetStatesAfterMacroDisabled();
-        hasCopper = true;
-        if (mc.currentScreen != null)
-            mc.thePlayer.closeScreen();
-        if (MacroHandler.getInstance().isMacroToggled())
-            Multithreading.schedule(() -> MacroHandler.getInstance().resumeMacro(), 1_500, TimeUnit.MILLISECONDS);
-        running = false;
+        if (!this.enabled) return;
+
+        this.enabled = false;
+        this.counter = 0;
+        this.resetStatesAfterMacroDisabled();
+        if (MacroHandler.getInstance().isMacroToggled()) {
+            MacroHandler.getInstance().resumeMacro();
+        }
+
+        log("Disabling");
     }
 
     @Override
     public void resetStatesAfterMacroDisabled() {
-        sprayState = AUTO_SPRAYONATOR_STATE.NONE;
-        running = false;
-        if (!sprayonatorPlotStates.isEmpty()) {
-            sprayState = AUTO_SPRAYONATOR_STATE.WAITING_FOR_PLOT;
-        }
-        checkPlotState = CHECK_PLOT_STATE.NONE;
-        skymartPurchaseState = SKYMART_PURCHASE_STATE.NONE;
-        bazaarPurchaseState = BAZAAR_PURCHASE_STATE.NONE;
-        currentGuiState = CURRENT_GUI_STATE.NONE;
+        this.mainState = MainState.NONE;
+        this.checkState = CheckPlotState.STARTING;
+        this.buyState = BuyState.STARTING;
+        this.sprayState = SprayState.STARTING;
     }
 
     @Override
-    public boolean isToggled() {
-        return FarmHelperConfig.enableSprayonator;
+    public boolean isToggled() { // more like should start
+        return FarmHelperConfig.autoSprayonatorEnable;
     }
 
     @Override
@@ -135,538 +120,353 @@ public class AutoSprayonator implements IFeature {
         return false;
     }
 
-    private boolean hasSprayonator() {
-        return InventoryUtils.hasItemInInventory("Sprayonator");
-    }
-
-    private int getChosenSprayonatorSlot() {
-        return FarmHelperConfig.sprayonatorSlot - 1;
-    }
-
-    private int getCurrentSprayonatorSlot() {
-        return InventoryUtils.getSlotIdOfItemInInventory("Sprayonator");
-    }
-
-    private boolean hasSprayItem() {
-        return InventoryUtils.hasItemInInventory(sprayItem.getItemName());
-    }
-
-    private boolean shouldBuySprayItem() {
-        return !FarmHelperConfig.sprayonatorItemInventoryOnly && !hasSprayItem();
-    }
-
-    private boolean sprayonatorInHotbar() {
-        return InventoryUtils.hasItemInHotbar("Sprayonator");
-    }
-
-    private boolean isHoldingSprayonator() {
-        return mc.thePlayer.inventory.currentItem == getChosenSprayonatorSlot();
-    }
-
-    private void moveSprayonator() {
-        if (getCurrentSprayonatorSlot() == getChosenSprayonatorSlot()) return;
-        if (getCurrentSprayonatorSlot() == -1) return;
-        if (mc.currentScreen == null) return;
-        if (mc.thePlayer.inventoryContainer.getSlot(getCurrentSprayonatorSlot()).getHasStack()) {
-            GuiInventory gui = (GuiInventory) mc.currentScreen;
-            mc.playerController.windowClick(gui.inventorySlots.windowId, getCurrentSprayonatorSlot(), getChosenSprayonatorSlot(), 2, this.mc.thePlayer);
-            Multithreading.schedule(mc.thePlayer::closeScreen, 100 + (long) (Math.random() * 50), TimeUnit.MILLISECONDS);
-            sprayState = AUTO_SPRAYONATOR_STATE.SET_SPRAYONATOR;
-        }
-    }
-
-    public void resetPlots() {
-        sprayonatorPlotStates.clear();
-        if (sprayState == AUTO_SPRAYONATOR_STATE.WAITING_FOR_PLOT) {
-            sprayState = AUTO_SPRAYONATOR_STATE.NONE;
-        }
-    }
-
+    // Thanks Xai <3
     @SubscribeEvent
     public void onTickShouldEnable(TickEvent.ClientTickEvent event) {
         if (mc.thePlayer == null || mc.theWorld == null) return;
-        if (running) return;
-        if (!isToggled()) return;
+        if (this.enabled) return;
+        if (!this.isToggled()) return;
         if (!MacroHandler.getInstance().isMacroToggled()) return;
         if (GameStateHandler.getInstance().getServerClosingSeconds().isPresent()) return;
+        if (MacroHandler.getInstance().isTeleporting()) return;
         if (FeatureManager.getInstance().isAnyOtherFeatureEnabled(this)) return;
         if (!GameStateHandler.getInstance().inGarden()) return;
-        if (sprayState != AUTO_SPRAYONATOR_STATE.WAITING_FOR_PLOT) return;
-        sprayItem = SPRAYONATOR_ITEM.values()[FarmHelperConfig.sprayonatorType];
+        if (this.mainState != MainState.NONE) return;
 
-        if (!hasSprayonator() && GameStateHandler.getInstance().getCopper() < 25) {
-            LogUtils.sendError("[Auto Sprayonator] Disabling due to no sprayonator");
-            return;
-        }
-        PlotData data = sprayonatorPlotStates.get(GameStateHandler.getInstance().getCurrentPlot());
-        if (data == null)
-            return;
-        if (sprayonatorPlotStates.isEmpty() || !data.isSprayed()) {
-            LogUtils.sendWarning("[Auto Sprayonator] Activating!");
-            LogUtils.sendDebug("[Auto Sprayonator] Plot Data: " + data);
-            KeyBindUtils.stopMovement();
-            start();
+
+        Plot currentPlot = this.plots[GameStateHandler.getInstance().getCurrentPlot()];
+        if (currentPlot == null || currentPlot.canSpray()) this.counter++;
+        else this.counter = 0;
+        if (this.counter > 150) {
+            this.start();
         }
     }
 
     @SubscribeEvent
-    public void onClientTick(TickEvent.ClientTickEvent event) {
-        if (mc.thePlayer == null || mc.theWorld == null) return;
-        if (!isToggled()) return;
-        if (FeatureManager.getInstance().isAnyOtherFeatureEnabled(this)) return;
-        if (event.phase != TickEvent.Phase.START) return;
-        if (!GameStateHandler.getInstance().inGarden()) return;
-        if (!MacroHandler.getInstance().isMacroToggled()) return;
-        if (sprayState != AUTO_SPRAYONATOR_STATE.NONE && sprayState != AUTO_SPRAYONATOR_STATE.WAITING_FOR_PLOT && !MacroHandler.getInstance().isCurrentMacroPaused()) {
-            MacroHandler.getInstance().pauseMacro();
-            return;
-        }
-        if (currentGuiState != CURRENT_GUI_STATE.NONE) return;
-        if (!sprayonatorDelay.passed()) return;
-        switch (sprayState) {
-            case CHECK_PLOTS:
-                switch (checkPlotState) {
-                    case OPEN_DESK:
-                        long delay = 100 + (long) (Math.random() * 100);
-                        sprayonatorDelay.schedule(500 + delay);
-                        if (mc.currentScreen != null && mc.currentScreen instanceof GuiChest) {
-                            // close gui and wait
-                            Multithreading.schedule(() -> {
-                                mc.thePlayer.closeScreen();
-                            }, delay, TimeUnit.MILLISECONDS);
-                            return;
-                        }
-                        currentGuiState = CURRENT_GUI_STATE.AWAIT_OPEN;
-                        Multithreading.schedule(() -> {
-                            mc.thePlayer.sendChatMessage("/desk");
-                        }, delay / 2, TimeUnit.MILLISECONDS);
-                        break;
-                    case END:
-                        checkPlotState = CHECK_PLOT_STATE.NONE;
-                        sprayState = AUTO_SPRAYONATOR_STATE.WAITING_FOR_PLOT;
-                        break;
-                    case NONE:
-                        checkPlotState = CHECK_PLOT_STATE.OPEN_DESK;
-                        break;
-                    default:
-                        break;
-                }
-                break;
-            case WAITING_FOR_PLOT:
-                if (sprayonatorPlotStates.isEmpty()) {
-                    sprayState = AUTO_SPRAYONATOR_STATE.CHECK_PLOTS;
-                    return;
-                }
-                if (!hasSprayItem() && !shouldBuySprayItem())
-                {
-                    stop();
-                    LogUtils.sendError("[Auto Sprayonator] Disabling until restart due to no spray item");
-                    sprayState = AUTO_SPRAYONATOR_STATE.NONE;
-                }
-                PlotData data = sprayonatorPlotStates.get(GameStateHandler.getInstance().getCurrentPlot());
-                if (data == null)
-                    return;
-                if (!data.isSprayed() || !data.sprayItem.equals(sprayItem.getItemName())) {
-                    sprayState = AUTO_SPRAYONATOR_STATE.CHECK_SPRAYONATOR;
-                    return;
-                } else {
-                    if (running && MacroHandler.getInstance().isCurrentMacroPaused()) {
-                        MacroHandler.getInstance().resumeMacro();
-                        stop();
-                    }
-                }
-                break;
-            case CHECK_SPRAYONATOR:
-                if (hasSprayonator()) {
-                    sprayState = AUTO_SPRAYONATOR_STATE.HOLD_SPRAYONATOR;
-                } else {
-                    sprayState = AUTO_SPRAYONATOR_STATE.SKYMART_PURCHASE;
-                }
-                return;
-            case SKYMART_PURCHASE:
-                switch (skymartPurchaseState) {
-                    case OPEN_DESK:
-                        break;
-                    case OPEN_SKYMART:
-                        break;
-                    case PURCHASE_ITEM:
-                        break;
-                    case END:
-                        break;
-                    case NONE:
-                        break;
-                }
-                break;
-            case HOLD_SPRAYONATOR:
-                sprayonatorDelay.schedule(2000);
-                if (sprayonatorInHotbar() && InventoryUtils.getSlotIdOfItemInHotbar("Sprayonator") == getChosenSprayonatorSlot()) {
-                    sprayState = AUTO_SPRAYONATOR_STATE.SET_SPRAYONATOR;
-                } else {
-                    if (mc.currentScreen != null) {
-                        Multithreading.schedule(() -> {
-                            mc.thePlayer.closeScreen();
-                        }, 100 + (long) (Math.random() * 50), TimeUnit.MILLISECONDS);
-                        return;
-                    }
-                    InventoryUtils.openInventory();
-                    Multithreading.schedule(this::moveSprayonator, 400 + (long) (Math.random() * 100), TimeUnit.MILLISECONDS);
-                }
-                break;
-            case SET_SPRAYONATOR:
-                if (mc.currentScreen != null) {
-                    Multithreading.schedule(() -> {
-                        mc.thePlayer.closeScreen();
-                    }, 100 + (long) (Math.random() * 50), TimeUnit.MILLISECONDS);
-                    return;
-                }
-                sprayonatorDelay.schedule(1000);
-                if (mc.thePlayer.inventory.currentItem != getChosenSprayonatorSlot()) {
-                    mc.thePlayer.inventory.currentItem = getChosenSprayonatorSlot();
-                }
-                List<String> itemLore = InventoryUtils.getItemLore(mc.thePlayer.inventory.getStackInSlot(getChosenSprayonatorSlot()));
-                if (itemLore.isEmpty()) return;
-                if (itemLore.size() < 19) return;
-                String sprayTypeLine = itemLore.get(18);
-                String sprayType = StringUtils.stripControlCodes(sprayTypeLine.split(": ")[1]);
-                if (sprayType.equals(sprayItem.getItemName())) {
-                    sprayState = AUTO_SPRAYONATOR_STATE.CHECK_ITEM;
-                    if (rotationState == 1) {
-                        RotationHandler.getInstance().easeTo(
-                                new RotationConfiguration(
-                                        new Rotation(prevRotation[0], prevRotation[1]),
-                                        FarmHelperConfig.getRandomRotationTime(), () -> {
-                                            rotationState = 0;
-                                            RotationHandler.getInstance().reset();
-                                        }
-                                ).easeOutBack(true)
-                        );
-                    }
-                } else {
-                    switch (rotationState) {
-                        case 0: // save rotations and rotate
-                            prevRotation = new float[]{
-                                    mc.thePlayer.rotationYaw,
-                                    mc.thePlayer.rotationPitch
-                            };
-                            RotationHandler.getInstance().easeTo(
-                                    new RotationConfiguration(
-                                            new Rotation(mc.thePlayer.rotationYaw, 77),
-                                            FarmHelperConfig.getRandomRotationTime(), null
-                                    ).easeOutBack(true)
-                            );
-                            rotationState++;
-                            break;
-                        case 1:
-                            Multithreading.schedule(KeyBindUtils::leftClick, 100 + (long) (Math.random() * 100), TimeUnit.MILLISECONDS);
-                            break;
-                    }
-
-                }
-                break;
-            case CHECK_ITEM:
-                if (hasSprayItem()) {
-                    sprayState = AUTO_SPRAYONATOR_STATE.USE_SPRAYONATOR;
-                } else {
-                    if (shouldBuySprayItem()) {
-                        sprayState = AUTO_SPRAYONATOR_STATE.BAZAAR_PURCHASE;
-                        bazaarPurchaseState = BAZAAR_PURCHASE_STATE.OPEN_BAZAAR;
-                    } else {
-                        stop();
-                        LogUtils.sendError("[Auto Sprayonator] Disabling until restart due to no spray item");
-                        sprayState = AUTO_SPRAYONATOR_STATE.NONE;
-                    }
-                }
-                break;
-            case BAZAAR_PURCHASE:
-                long randomDelay = (long) (Math.random() * 250);
-                switch (bazaarPurchaseState) {
-                    case OPEN_BAZAAR:
-                        mc.thePlayer.sendChatMessage("/bz " + sprayItem.getItemName().toLowerCase());
-                        bazaarItemName = sprayItem.getItemName();
-                        sprayonatorDelay.schedule(1000 + randomDelay);
-                        currentGuiState = CURRENT_GUI_STATE.AWAIT_OPEN;
-                        break;
-                    case CLICK_ITEM: // handled in gui event
-                    case BUY_ITEM:
-                        break;
-                    case NONE:
-                        break;
-                }
-                break;
-            case USE_SPRAYONATOR:
-                if (isHoldingSprayonator()) {
-                    sprayonatorDelay.schedule(1000);
-                    Multithreading.schedule(() -> {
-                        KeyBindUtils.rightClick();
-                        sprayState = AUTO_SPRAYONATOR_STATE.WAITING_FOR_PLOT;
-                    }, 100 + (long) (Math.random() * 50), TimeUnit.MILLISECONDS);
-                }
-                break;
+    public void onTickMainState(TickEvent.ClientTickEvent event) {
+        if (mc.thePlayer == null || mc.theWorld == null || !this.enabled) return;
+        switch (this.mainState) {
             case NONE:
                 break;
-        }
-    }
+            case STARTING:
+                // Remove these varialbes if u need
+                Plot currentPlot = this.plots[GameStateHandler.getInstance().getCurrentPlot()];
+                boolean shouldCheckPlots = Arrays.stream(this.plots).noneMatch(Objects::nonNull);
+                boolean sprayonatorItemNotInInventory = !InventoryUtils.hasItemInInventory(this.SPRAYONATOR_ITEM[FarmHelperConfig.autoSprayonatorType]);
+                boolean canSpray = currentPlot != null && currentPlot.canSpray();
 
-    @SubscribeEvent
-    public void onGuiOpen(DrawScreenAfterEvent event) {
-        if (mc.thePlayer == null || mc.theWorld == null) return;
-        if (!sprayonatorDelay.passed()) return;
-        if (!GameStateHandler.getInstance().inGarden()) return;
-        if (!running) return;
-        if (!(event.guiScreen instanceof GuiChest)) return;
-        String guiName = InventoryUtils.getInventoryName();
-        if (guiName == null) return;
-        ContainerChest guiChest = (ContainerChest) ((GuiChest) event.guiScreen).inventorySlots;
-        if (currentGuiState != CURRENT_GUI_STATE.AWAIT_OPEN) return;
-        sprayItem = SPRAYONATOR_ITEM.values()[FarmHelperConfig.sprayonatorType];
-
-        long randomDelay = (long) (Math.random() * 250);
-
-        switch (sprayState) {
-            case CHECK_PLOTS:
-                sprayonatorDelay.schedule(FarmHelperConfig.sprayonatorAdditionalDelay + 1000 + randomDelay);
-                switch (checkPlotState) {
-                    case OPEN_DESK:
-                        if (!guiName.equals("Desk")) {
-                            currentGuiState = CURRENT_GUI_STATE.CLOSING;
-                            Multithreading.schedule(() -> {
-                                mc.thePlayer.closeScreen();
-                                currentGuiState = CURRENT_GUI_STATE.NONE;
-                            }, FarmHelperConfig.sprayonatorAdditionalDelay + 100 + (long) (Math.random() * 50), TimeUnit.MILLISECONDS);
-                            return;
-                        }
-                        checkPlotState = CHECK_PLOT_STATE.OPEN_PLOTS;
-                        Multithreading.schedule(() -> {
-                            InventoryUtils.clickContainerSlot(InventoryUtils.getSlotIdOfItemInContainer("Configure Plots"), InventoryUtils.ClickType.LEFT, InventoryUtils.ClickMode.PICKUP);
-                        }, FarmHelperConfig.sprayonatorAdditionalDelay + 100 + (long) (Math.random() * 50), TimeUnit.MILLISECONDS);
-                        break;
-                    case OPEN_PLOTS:
-                        if (!guiName.equals("Configure Plots")) {
-                            currentGuiState = CURRENT_GUI_STATE.CLOSING;
-                            Multithreading.schedule(() -> {
-                                mc.thePlayer.closeScreen();
-                                currentGuiState = CURRENT_GUI_STATE.NONE;
-                            }, FarmHelperConfig.sprayonatorAdditionalDelay + 100 + (long) (Math.random() * 50), TimeUnit.MILLISECONDS);
-                            return;
-                        }
-                        loadSprayonatorData(guiChest);
-                        currentGuiState = CURRENT_GUI_STATE.CLOSING;
-                        checkPlotState = CHECK_PLOT_STATE.END;
-                        sprayState = AUTO_SPRAYONATOR_STATE.NONE;
-                        Multithreading.schedule(() -> {
-                            mc.thePlayer.closeScreen();
-                            sprayState = AUTO_SPRAYONATOR_STATE.WAITING_FOR_PLOT;
-                            checkPlotState = CHECK_PLOT_STATE.NONE;
-                            currentGuiState = CURRENT_GUI_STATE.NONE;
-                        }, FarmHelperConfig.sprayonatorAdditionalDelay + 150 + (long) (Math.random() * 50), TimeUnit.MILLISECONDS);
-                }
-                break;
-            case BAZAAR_PURCHASE:
-                sprayonatorDelay.schedule(FarmHelperConfig.sprayonatorAdditionalDelay + 2000 + randomDelay * 3);
-                if (guiName.contains("Bazaar")) {
-                    if (bazaarPurchaseState == BAZAAR_PURCHASE_STATE.OPEN_BAZAAR)
-                        bazaarPurchaseState = BAZAAR_PURCHASE_STATE.CLICK_ITEM;
-                } else {
-                    bazaarPurchaseState = BAZAAR_PURCHASE_STATE.NONE;
-                    currentGuiState = CURRENT_GUI_STATE.CLOSING;
-                    Multithreading.schedule(() -> {
-                        mc.thePlayer.closeScreen();
-                        currentGuiState = CURRENT_GUI_STATE.NONE;
-                    }, FarmHelperConfig.sprayonatorAdditionalDelay + 400 + (long) (Math.random() * 50), TimeUnit.MILLISECONDS);
-                    return;
-                }
-                LogUtils.sendDebug("Bazaar Purchase State: " + bazaarPurchaseState.toString());
-                switch (bazaarPurchaseState) {
-                    case CLICK_ITEM:
-                        Multithreading.schedule(() -> {
-                            Slot slot = InventoryUtils.getSlotOfItemInContainer(bazaarItemName, true);
-                            LogUtils.sendDebug("Item Slot: " + (slot == null ? "null" : slot.slotNumber));
-                            if (slot == null) {
-                                currentGuiState = CURRENT_GUI_STATE.CLOSING;
-                                Multithreading.schedule(() -> {
-                                    mc.thePlayer.closeScreen();
-                                    currentGuiState = CURRENT_GUI_STATE.NONE;
-                                }, FarmHelperConfig.sprayonatorAdditionalDelay + 100 + (long) (Math.random() * 50), TimeUnit.MILLISECONDS);
-                                return;
-                            }
-                            InventoryUtils.clickContainerSlot(slot.slotNumber, InventoryUtils.ClickType.LEFT, InventoryUtils.ClickMode.PICKUP);
-                            bazaarPurchaseState = BAZAAR_PURCHASE_STATE.BUY_ITEM;
-                            sprayonatorDelay.schedule(-1);
-                        }, FarmHelperConfig.sprayonatorAdditionalDelay + 300 + (long) (Math.random() * 50), TimeUnit.MILLISECONDS);
-                        break;
-                    case BUY_ITEM:
-                        Multithreading.schedule(() -> {
-                            LogUtils.sendDebug("Buy Item");
-                            if (guiName.contains(bazaarItemName) || guiName.contains(bazaarItemName.toLowerCase())) {
-                                Slot slot = InventoryUtils.getSlotOfItemInContainer("Buy Instantly", false);
-                                LogUtils.sendDebug("Buy Slot: " + (slot == null ? "null" : slot.slotNumber));
-                                if (slot == null) {
-                                    currentGuiState = CURRENT_GUI_STATE.CLOSING;
-                                    Multithreading.schedule(() -> {
-                                        mc.thePlayer.closeScreen();
-                                        currentGuiState = CURRENT_GUI_STATE.NONE;
-                                    }, FarmHelperConfig.sprayonatorAdditionalDelay + 250 + (long) (Math.random() * 50), TimeUnit.MILLISECONDS);
-                                    return;
-                                }
-                                Multithreading.schedule(() -> {
-                                    InventoryUtils.clickContainerSlot(slot.slotNumber, InventoryUtils.ClickType.LEFT, InventoryUtils.ClickMode.PICKUP);
-                                    Multithreading.schedule(() -> {
-                                        InventoryUtils.clickContainerSlot(slot.slotNumber, InventoryUtils.ClickType.LEFT, InventoryUtils.ClickMode.PICKUP);
-                                        currentGuiState = CURRENT_GUI_STATE.CLOSING;
-                                        Multithreading.schedule(() -> {
-                                            mc.thePlayer.closeScreen();
-                                            currentGuiState = CURRENT_GUI_STATE.NONE;
-                                            sprayState = AUTO_SPRAYONATOR_STATE.CHECK_ITEM;
-                                        }, 300 + (long) (Math.random() * 50), TimeUnit.MILLISECONDS);
-                                    }, FarmHelperConfig.sprayonatorAdditionalDelay/2 + 300 + (long) (Math.random() * 50), TimeUnit.MILLISECONDS);
-                                }, FarmHelperConfig.sprayonatorAdditionalDelay/2 + 300 + (long) (Math.random() * 50), TimeUnit.MILLISECONDS);
-                                bazaarPurchaseState = BAZAAR_PURCHASE_STATE.NONE;
-                            } else {
-                                LogUtils.sendDebug("Invalid GUI: " + guiName);
-                                currentGuiState = CURRENT_GUI_STATE.CLOSING;
-                                Multithreading.schedule(() -> {
-                                    mc.thePlayer.closeScreen();
-                                    currentGuiState = CURRENT_GUI_STATE.NONE;
-                                }, FarmHelperConfig.sprayonatorAdditionalDelay + 100 + (long) (Math.random() * 50), TimeUnit.MILLISECONDS);
-                            }
-                        }, FarmHelperConfig.sprayonatorAdditionalDelay + 300 + (long) (Math.random() * 50), TimeUnit.MILLISECONDS);
-                        break;
-                    default:
-                        break;
-                }
-            default:
-                break;
-        }
-    }
-
-    @SubscribeEvent
-    public void onChatReceived(ClientChatReceivedEvent e) {
-        if (mc.thePlayer == null || mc.theWorld == null) return;
-        if (!running) return;
-        if (!isToggled()) return;
-        if (e.type != 0) return;
-        if (e.message.getUnformattedText().contains("You sprayed Plot")) {
-            String plotNumber = e.message.getUnformattedText().split(" ")[5];
-            PlotData data = new PlotData(Integer.parseInt(plotNumber), sprayItem.getItemName(), TimeUnit.MINUTES.toMillis(30));
-            sprayonatorPlotStates.put(Integer.parseInt(plotNumber), data);
-        } else if (e.message.getUnformattedText().contains("sprayed with that item recently")) {
-            sprayState = AUTO_SPRAYONATOR_STATE.CHECK_PLOTS;
-        }
-    }
-
-    private void loadSprayonatorData(ContainerChest guiChest) {
-        int plotCounter = 0;
-        for (int i = 0; i < guiChest.inventorySlots.size(); i++) {
-            Slot slot = guiChest.inventorySlots.get(i);
-            if (slot == null || !slot.getHasStack()) continue;
-            if (slot.getStack().getDisplayName().contains("Plot")) {
-                int plotNumber = PlotUtils.getPLOT_NUMBERS().get(plotCounter);
-                List<String> lore = InventoryUtils.getItemLore(slot.getStack());
-                boolean foundSpray = false;
-                for (String line : lore) {
-                    if (line.contains("Sprayed with")) {
-                        System.out.println(line);
-                        Matcher matcher = sprayTimerPattern.matcher(line.replace("Sprayed with ", ""));
-                        if (matcher.find()) {
-                            String minutes = matcher.group(2);
-                            if (minutes == null) {
-                                minutes = "0";
-                            }
-                            String seconds = matcher.group(3);
-                            long time = TimeUnit.MINUTES.toMillis(Integer.parseInt(minutes))
-                                    + TimeUnit.SECONDS.toMillis(Integer.parseInt(seconds));
-                            PlotData data = new PlotData(plotNumber, matcher.group(1).trim(), time);
-                            foundSpray = true;
-                            sprayonatorPlotStates.put(plotNumber, data);
-                        }
-                        break;
+                if (shouldCheckPlots) {
+                    this.mainState = MainState.CHECK_PLOTS;
+                    this.checkState = CheckPlotState.STARTING;
+                } else if (sprayonatorItemNotInInventory) {
+                    log("Cannot find sprayonator item in inventory.");
+                    if (FarmHelperConfig.autoSprayonatorAutoBuyItem) {
+                        this.mainState = MainState.BUY_ITEM;
+                        this.buyState = BuyState.STARTING;
+                    } else {
+                        FarmHelperConfig.autoSprayonatorEnable = false;
+                        this.stop();
+                        return;
                     }
+                } else if (canSpray) {
+                    this.timer.schedule(FarmHelperConfig.autoSprayonatorAdditionalDelay);
+                    this.mainState = MainState.SPRAY_PLOT;
+                    this.sprayState = SprayState.STARTING;
+                } else {
+                    log("Cannot spray/Plot is already sprayed. Disabling");
+                    this.stop();
                 }
-                if (!foundSpray) {
-                    PlotData data = new PlotData(plotNumber);
-                    sprayonatorPlotStates.put(plotNumber, data);
-                }
-                plotCounter++;
-            } else if (slot.getStack().getDisplayName().contains("Barn")) {
-                plotCounter++;
+                break;
+            case CHECK_PLOTS:
+                this.handleCheckPlots();
+                break;
+            case BUY_ITEM:
+                this.handleBuyItem();
+                break;
+            case SPRAY_PLOT:
+                this.handleSprayPlot();
+                break;
+        }
+    }
+
+    @SubscribeEvent
+    public void onChatReceive(ClientChatReceivedEvent event) {
+        if (event.type != 0 || !this.enabled) return;
+        String message = StringUtils.stripControlCodes(event.message.getUnformattedText());
+        if (message.startsWith("SPRAYONATOR! Your selected material is now ") && this.sprayState == SprayState.SPRAYONATOR_TYPE_VERIFY) {
+            log("Sprayonator Material Changed");
+            String itemName = SPRAYONATOR_ITEM[FarmHelperConfig.autoSprayonatorType];
+            this.changedMaterial = true;
+            if (message.contains(itemName)) {
+                this.changedMaterial = false;
+                this.sprayState = SprayState.STARTING;
             }
         }
-        sprayonatorPlotStates.forEach((k, v) -> {
-            LogUtils.sendDebug("Plot " + k + " : " + v.toString());
-        });
+        if (message.startsWith("SPRAYONATOR! You sprayed Plot - ") && this.sprayState == SprayState.SPRAY_VERIFY) {
+            this.plots[GameStateHandler.getInstance().getCurrentPlot()].setSprayExpireTime(System.currentTimeMillis() + 1800000); // 30 minutes
+            this.sprayState = SprayState.DISABLE;
+            this.timer.schedule(FarmHelperConfig.autoSprayonatorAdditionalDelay);
+        }
+        if (message.contains("This plot was sprayed with that item recently! Try again soon!")) {
+            this.plots = new Plot[25];
+            this.mainState = MainState.STARTING;
+            this.checkState = CheckPlotState.STARTING;
+        }
     }
 
-    enum AUTO_SPRAYONATOR_STATE {
-        CHECK_PLOTS, // save states of all plots
-        WAITING_FOR_PLOT, // waiting to enter a non sprayed plot
-        CHECK_SPRAYONATOR, // check if sprayonator in inventory
-        SKYMART_PURCHASE, // buy sprayonator if not in inventory
-        HOLD_SPRAYONATOR, // move sprayonator into inventory slot
-        SET_SPRAYONATOR, // set sprayonator item to correct item
-        CHECK_ITEM, // check if item is in inventory
-        BAZAAR_PURCHASE, // buy item if not in inventory
-        USE_SPRAYONATOR, // use sprayonator on current plot
-        NONE,
+    private void handleCheckPlots() {
+        switch (this.checkState) {
+            case STARTING:
+                this.checkState = CheckPlotState.OPEN_DESK;
+                break;
+            case OPEN_DESK:
+                mc.thePlayer.sendChatMessage("/desk");
+                this.checkState = CheckPlotState.VERIFY_DESK;
+                this.timer.schedule(1000);
+                break;
+            case VERIFY_DESK:
+                if (this.timer.isScheduled() && this.timer.passed()) {
+                    log("Could not open desk. Disabling");
+                    this.stop();
+                    return;
+                }
+                // WHy null? Why not ""? so that i can use contains without null issues??? why man?
+                String inventoryName = InventoryUtils.getInventoryName();
+                if ((mc.currentScreen instanceof GuiChest || mc.thePlayer.openContainer instanceof ContainerChest)
+                    && (inventoryName != null && inventoryName.contains("Desk"))) {
+                    log("Opened Desk.");
+                    this.checkState = CheckPlotState.OPEN_PLOTS;
+                    this.timer.schedule(FarmHelperConfig.autoSprayonatorAdditionalDelay);
+                }
+                break;
+            case OPEN_PLOTS:
+                if (!this.hasTimerEnded()) return;
+                InventoryUtils.clickContainerSlot(InventoryUtils.getSlotIdOfItemInContainer("Configure Plots"), InventoryUtils.ClickType.LEFT, InventoryUtils.ClickMode.PICKUP);
+                this.checkState = CheckPlotState.VERIFY_PLOTS;
+                this.timer.schedule(FarmHelperConfig.autoSprayonatorAdditionalDelay);
+                log("Clicked on Configure Plots Hopefully");
+                break;
+            case VERIFY_PLOTS:
+                if (this.timer.isScheduled() && this.timer.passed()) {
+                    log("Could not open plots. Disabling");
+                    this.stop();
+                    return;
+                }
+                String inventoryName2 = InventoryUtils.getInventoryName();
+                if ((mc.currentScreen instanceof GuiChest || mc.thePlayer.openContainer instanceof ContainerChest)
+                    && (inventoryName2 != null && inventoryName2.contains("Configure Plots"))) {
+                    log("In Configure Plots Menu");
+                    this.checkState = CheckPlotState.SCAN_PLOTS;
+                    this.timer.schedule(FarmHelperConfig.autoSprayonatorAdditionalDelay);
+                }
+                break;
+            case SCAN_PLOTS:
+                if (!this.hasTimerEnded()) return;
+                int currentPlotIndex = 0;
+                log("Scanning Plots");
+                for (int i = 0; i < mc.thePlayer.openContainer.inventorySlots.size() - 37; i++) {
+                    Slot slot = mc.thePlayer.openContainer.getSlot(i);
+                    ItemStack slotStack = slot.getStack();
+                    if (slotStack == null) continue;
+                    if (slotStack.getDisplayName().contains("The Barn")) {
+                        currentPlotIndex++;
+                        continue;
+                    }
+                    if (!slotStack.getDisplayName().contains("Plot")) continue;
+
+                    int actualPlotNumber = PlotUtils.getPLOT_NUMBERS().get(currentPlotIndex);
+                    int minute = 0;
+                    int second = 0;
+                    String lore = String.join(" ", InventoryUtils.getLoreOfItemInContainer(i));
+                    Matcher timeMatcher = this.sprayonatorTimePattern.matcher(lore);
+                    if (lore.contains("Sprayed with ") && timeMatcher.find()) {
+                        second = Integer.parseInt(timeMatcher.group(timeMatcher.groupCount()));
+                        String minuteString = timeMatcher.group(timeMatcher.groupCount() - 1);
+                        if (minuteString != null) {
+                            minute = Integer.parseInt(minuteString);
+                        }
+                    }
+                    long sprayExpiryTime = System.currentTimeMillis() + (minute * 60000L) + (second * 1000L);
+                    this.plots[actualPlotNumber] = new Plot(actualPlotNumber, sprayExpiryTime);
+                    currentPlotIndex++;
+                }
+                log(Arrays.toString(this.plots));
+                this.timer.schedule(FarmHelperConfig.autoSprayonatorAdditionalDelay);
+                this.checkState = CheckPlotState.DISABLE;
+                break;
+            case DISABLE:
+                if (!this.hasTimerEnded()) return;
+                this.checkState = CheckPlotState.STARTING;
+                this.mainState = MainState.STARTING;
+                PlayerUtils.closeScreen();
+                break;
+        }
     }
 
-    enum CHECK_PLOT_STATE {
-        OPEN_DESK,
-        OPEN_PLOTS,
-        END,
-        NONE
+    private void handleBuyItem() {
+        switch (this.buyState) {
+            case STARTING:
+                this.buyState = BuyState.TOGGLE_AUTO_BZ;
+                break;
+            case TOGGLE_AUTO_BZ:
+                AutoBazaar.getInstance().buy(SPRAYONATOR_ITEM[FarmHelperConfig.autoSprayonatorType], FarmHelperConfig.autoSprayonatorAutoBuyAmount, 0);
+                this.buyState = BuyState.VERIFY_AUTO_BZ;
+                break;
+            case VERIFY_AUTO_BZ:
+                if (AutoBazaar.getInstance().hasFailed()) {
+                    LogUtils.sendError("[AutoSprayonator] - Failed to buy item from bazaar. Stopping AutoSprayonator");
+                    FarmHelperConfig.autoSprayonatorEnable = false;
+                    this.stop();
+                    return;
+                }
+                if (AutoBazaar.getInstance().hasSucceeded()) {
+                    log("Bought item from bazaar.");
+                    this.buyState = BuyState.DISABLE;
+                }
+                break;
+            case DISABLE:
+                this.buyState = BuyState.STARTING;
+                this.mainState = MainState.STARTING;
+                break;
+        }
     }
 
-    enum SKYMART_PURCHASE_STATE {
-        OPEN_DESK,
-        OPEN_SKYMART,
-        PURCHASE_ITEM,
-        END,
-        NONE
+    private void handleSprayPlot() {
+        switch (sprayState) {
+            case STARTING:
+                if (!this.hasTimerEnded()) return;
+
+                PlayerUtils.closeScreen();
+
+                this.sprayState = SprayState.USE_SPRAYONATOR;
+
+                // Not necessary but just in case
+                if (!InventoryUtils.hasItemInInventory(SPRAYONATOR)) {
+                    this.stop();
+                    LogUtils.sendError("Sprayonator not in inventory.");
+                    return;
+                }
+                if (!InventoryUtils.hasItemInHotbar(SPRAYONATOR)) {
+                    InventoryUtils.openInventory();
+                    this.sprayState = SprayState.SWAP_SPRAYONATOR_TO_HOTBAR;
+                    this.timer.schedule(FarmHelperConfig.autoSprayonatorAdditionalDelay);
+                    return;
+                }
+                if (!InventoryUtils.holdItem(SPRAYONATOR)) {
+                    this.stop();
+                    LogUtils.sendError("Cannot hold sprayonator.");
+                    return;
+                }
+                String sprayonatorLore = String.join(" ", InventoryUtils.getLoreOfItemInContainer(InventoryUtils.getSlotIdOfItemInContainer(SPRAYONATOR)));
+                if (!sprayonatorLore.contains(this.SPRAYONATOR_ITEM[FarmHelperConfig.autoSprayonatorType])) {
+                    this.sprayState = SprayState.CHANGE_SPRAYONATOR_TYPE;
+                    this.isRotating = true;
+
+                    RotationConfiguration rotationConfig = new RotationConfiguration(
+                        new Rotation(AngleUtils.getActualYawFrom360(mc.thePlayer.rotationYaw), -60),
+                        500,
+                        () -> {
+                            this.isRotating = false;
+                            RotationHandler.getInstance().reset();
+                        });
+                    RotationHandler.getInstance().easeTo(rotationConfig);
+
+                    return;
+                }
+                break;
+            case CHANGE_SPRAYONATOR_TYPE:
+                if (this.isRotating || !this.hasTimerEnded()) return;
+
+                KeyBindUtils.leftClick();
+                this.sprayState = SprayState.SPRAYONATOR_TYPE_VERIFY;
+                this.timer.schedule(2000);
+                break;
+            case SPRAYONATOR_TYPE_VERIFY:
+                if (this.hasTimerEnded()) {
+                    log("Cannot verify sprayonator type");
+                    this.stop();
+                    return;
+                }
+                if (!this.changedMaterial) return;
+
+                this.changedMaterial = false;
+                this.sprayState = SprayState.CHANGE_SPRAYONATOR_TYPE;
+                this.timer.schedule(500);
+                break;
+            case SWAP_SPRAYONATOR_TO_HOTBAR:
+                if (!this.hasTimerEnded()) return;
+
+                InventoryUtils.swapSlots(InventoryUtils.getSlotIdOfItemInInventory(SPRAYONATOR), FarmHelperConfig.autoSprayonatorSlot - 1);
+                this.timer.schedule(FarmHelperConfig.autoSprayonatorAdditionalDelay);
+                this.sprayState = SprayState.STARTING;
+                break;
+            case USE_SPRAYONATOR:
+                KeyBindUtils.rightClick();
+                this.sprayState = SprayState.SPRAY_VERIFY;
+                this.timer.schedule(2000);
+                break;
+            case SPRAY_VERIFY:
+                if (this.hasTimerEnded()) {
+                    LogUtils.sendError("Could not Spray Plot. Disabling.");
+                    this.stop();
+                    return;
+                }
+            case DISABLE:
+                if (!this.hasTimerEnded()) return;
+                log("Disabling spray handler.");
+                this.stop();
+                break;
+        }
     }
 
-    enum BAZAAR_PURCHASE_STATE {
-        OPEN_BAZAAR,
-        CLICK_ITEM,
-        BUY_ITEM,
-        NONE
+    private boolean hasTimerEnded() {
+        return (this.timer.isScheduled() && this.timer.passed());
     }
 
-    enum CURRENT_GUI_STATE {
-        AWAIT_OPEN,
-        CLOSING,
-        NONE
+    private void log(String message) {
+        LogUtils.sendDebug("[AutoSprayonator] - " + message);
     }
 
-    public static class PlotData {
-        @Getter
-        private final int plot_number;
-        @Getter
-        private final Clock sprayClock = new Clock();
+    public void resetPlots() {
+        this.plots = new Plot[25];
+    }
 
-        @Setter @Getter
-        private String sprayItem = "none";
-        @Setter
-        private long sprayTime;
+    enum MainState {
+        NONE, STARTING, CHECK_PLOTS, BUY_ITEM, SPRAY_PLOT
+    }
 
-        private PlotData(int plot_number) {
-            this.plot_number = plot_number;
-            sprayClock.schedule(0);
+    enum CheckPlotState {
+        STARTING, OPEN_DESK, VERIFY_DESK,
+        OPEN_PLOTS, VERIFY_PLOTS,
+        SCAN_PLOTS, DISABLE
+    }
+
+    enum BuyState {
+        STARTING, TOGGLE_AUTO_BZ, VERIFY_AUTO_BZ, DISABLE
+    }
+
+    enum SprayState {
+        STARTING, CHANGE_SPRAYONATOR_TYPE, SPRAYONATOR_TYPE_VERIFY, SWAP_SPRAYONATOR_TO_HOTBAR, USE_SPRAYONATOR, SPRAY_VERIFY, DISABLE
+    }
+
+    @Getter
+    @Setter
+    static class Plot {
+        private final int plotNumber;
+        private long sprayExpireTime = 0L;
+
+        public Plot(int plotNumber, long sprayExpireTime) {
+            this.plotNumber = plotNumber;
+            this.sprayExpireTime = sprayExpireTime;
         }
 
-        private PlotData(int plot_number, String spray_item, long time) {
-            this.plot_number = plot_number;
-            this.sprayItem = spray_item;
-            this.sprayTime = time;
-            sprayClock.schedule(time);
-        }
-
-        public boolean isSprayed() {
-            return !sprayClock.passed();
+        public boolean canSpray() {
+            return System.currentTimeMillis() > sprayExpireTime;
         }
 
         @Override
         public String toString() {
-            return "{" + "sprayed=" + isSprayed() + ", sprayItem='" + sprayItem + '\'' + ", sprayTime=" + sprayTime + ", plot_number=" + plot_number + '}';
+            return String.format("Plot[%d, canSpray = %b]", this.plotNumber, this.canSpray());
         }
     }
 }
